+++ src/EnhancedScatterChart.ts
@@ -352,7 +352,6 @@ export class EnhancedScatterChart implements IVisual {
     private formattingSettingsService: FormattingSettingsService;
 
     private hasHighlights: boolean;
-    private useImages: boolean;
 
     private legend: ILegend;
 
@@ -721,7 +720,7 @@ export class EnhancedScatterChart implements IVisual {
         visualHost: IVisualHost,
         interactivityService: IInteractivityService<BaseDataPoint>,
     ): EnhancedScatterChartData {
-        this.formattingSettings = this.formattingSettingsService.populateFormattingSettingsModel(EnhancedScatterChartSettingsModel, [dataView]);
+        this.formattingSettings = this.formattingSettingsService.populateFormattingSettingsModel(EnhancedScatterChartSettingsModel, dataView);
         const settings: EnhancedScatterChartSettingsModel = this.formattingSettings;
 
         this.parseSettings(new ColorHelper(colorPalette));
@@ -766,9 +765,6 @@ export class EnhancedScatterChart implements IVisual {
         }
 
         this.hasHighlights = dataValues.length > 0 && dataValues.some(value => value.highlights && value.highlights.some(_ => _));
-
-        /* Property is used for detection of image bucket in visual */
-        this.useImages = scatterMetadata.idx.image >= EnhancedScatterChart.MinIndex;
 
         const sizeRange: ValueRange<number> = EnhancedScatterChart.getSizeRangeForGroups(
             grouped,
@@ -1450,7 +1446,7 @@ export class EnhancedScatterChart implements IVisual {
                     }
                 }
 
-                const fillEnabled = settings.enableFillPointCardSettings.show.value && !this.useImages;
+                const fillEnabled = settings.enableFillPointCardSettings.show.value && !this.data?.useShape;
                 const outlineEnabled = settings.enableOutlineCardSettings.show.value;
                 const outlineWidth = settings.enableOutlineCardSettings.strokeWidth.value;
         
@@ -2877,7 +2873,7 @@ export class EnhancedScatterChart implements IVisual {
         let markers: Selection<EnhancedScatterChartDataPoint>,
             markersMerged: Selection<EnhancedScatterChartDataPoint>;
 
-        const markersChanged = this.data.useShape ? this.drawScatterMarkersUsingShapes(markers, markersMerged, scatterData, sizeRange, duration) :
+        const markersChanged = this.data?.useShape ? this.drawScatterMarkersUsingShapes(markers, markersMerged, scatterData, sizeRange, duration) :
                 this.drawScatterMarkersWithoutShapes(markers, markersMerged, scatterData, sizeRange, duration);
 
         const newMarkers: Selection<EnhancedScatterChartDataPoint> = markersChanged.markers,
@@ -3098,52 +3094,38 @@ export class EnhancedScatterChart implements IVisual {
 
         const settings: EnhancedScatterChartSettingsModel = this.formattingSettings;
 
-        const newCards = [...settings.cards];
-
         settings.cards.forEach(element => {
             switch (element.name) {
                 case "dataPoint": {
-                    if (this.data && this.data.hasGradientRole) {
-                        this.removeArrayItem(newCards, settings.enableDataPointCardSettings);
+                    if (this.data?.hasGradientRole || this.data?.useCustomColor) {
+                        settings.enableDataPointCardSettings.visible = false;
                     }
                     settings.populateColorSelector(this.data.legendDataPoints, this.data.dataPoints);
 
                     break;
                 }
                 case "fillPoint": {
-                    if (this.useImages) {
-                        this.removeArrayItem(newCards, settings.enableFillPointCardSettings);
+                    if (this.data?.useShape) {
+                        settings.enableFillPointCardSettings.visible = false;
                     }
 
                     break;
                 }
                 case "legend": {
                     if (!this.data || !this.data.hasDynamicSeries) {
-                        this.removeArrayItem(newCards, settings.enableLegendCardSettings);
+                        settings.enableLegendCardSettings.visible = false;
                     }
 
                     break;
                 }
                 case "outline": {
-                    if (this.useImages) {
-                        this.removeArrayItem(newCards, settings.enableOutlineCardSettings);
+                    if (this.data?.useShape) {
+                        settings.enableOutlineCardSettings.visible = false;
                     }
 
                     break;
                 }
             }
         });
-
-        settings.cards = newCards;
-        this.formattingSettings = settings;
-    }
-
-    private removeArrayItem<T>(array: T[], item: T)
-    {
-        const index: number = array.indexOf(item);
-        if (index > -1)
-        {
-            array.splice(index, 1);
-        }
     }
 }